KINETIC blocks working on cpu
CoreNEURON Repo SHA: BlueBrain/CoreNeuron@dbb612e86393c2c34f82b69bc55453292b666995

+++ src/coreneuron/nrnoc/nrnoc_aux.c
@@ -72,9 +72,8 @@ void freevector(double* p) {
 double** makematrix(size_t nrows, size_t ncols) {
     size_t i;
     double **matrix;
-    
-    matrix = (double **) emalloc((unsigned) (nrows * sizeof(double *)));
-    *matrix = (double *) emalloc((unsigned) (nrows * ncols * sizeof(double)));
+    matrix = (double **) emalloc(nrows * sizeof(double *));
+    *matrix = (double *) emalloc(nrows * ncols * sizeof(double));
     for (i = 1; i < nrows; i++)
         matrix[i] = matrix[i - 1] + ncols;
     return (matrix);
